Extract a file for the details only
